Fix destroying array

+++ CodeForces/722C_Destroying_Array.cpp
@@ -17,38 +17,38 @@ int main(){
     vector<int> destroy_indices(n+1);
     for(int i=1; i<=n; i++) cin >> arr[i];
     for(int i=1; i<=n; i++) cin >> destroy_indices[i];
-    vector<int> result;
+    vector<long long> result;
     result.push_back(0);
     vector<int> parents(n+1);
-    vector<int> sum(n+1);
+    vector<long long> sum(n+1);
     vector<int> temp_arr(n+1);
     fill(temp_arr.begin(), temp_arr.end(), -1);
     for(int i=1; i<=n; i++) parents[i] = i;
     fill(sum.begin(), sum.end(), -1);
-    int max = 0;
+    long long max = 0;
     for(int i=n; i>1; i--){
-        int new_index = destroy_indices[i];
-        int new_number = arr[new_index];
+        long long new_index = destroy_indices[i];
+        long long new_number = (long long) arr[new_index];
         temp_arr[new_index] = new_number;
-        int new_sum;
+        long long new_sum;
         if(new_index-1 > 0 && temp_arr[new_index-1] != -1 && new_index+1 <= n && temp_arr[new_index+1] != -1){
             int root1 = find_root(new_index-1, parents);
             int root2 = find_root(new_index+1, parents);
-            int sum1 = sum[root1];
-            int sum2 = sum[root2];
+            long long sum1 = sum[root1];
+            long long sum2 = sum[root2];
             new_sum = sum1 + new_number + sum2;
             sum[root1] = new_sum;
             parents[new_index] = root1;
             parents[root2] = root1;
         } else if (new_index-1 > 0 && temp_arr[new_index-1] != -1){
             int root1 = find_root(new_index-1, parents);
-            int sum1 = sum[root1];
+            long long sum1 = sum[root1];
             new_sum = sum1 + new_number;
             sum[root1] = new_sum;
             parents[new_index] = root1;
         } else if (new_index+1 <= n && temp_arr[new_index+1] != -1){
             int root2 = find_root(new_index+1, parents);
-            int sum2 = sum[root2];
+            long long sum2 = sum[root2];
             new_sum = sum2 + new_number;
             sum[root2] = new_sum;
             parents[new_index] = root2;
